move integration tests to test_plugin
