--- tests/assays/test_curiosity/test_integration_bradleyterryevaluator.readout.json
@@ -1,30 +1,0 @@
-{
-  "passed": true,
-  "details": {
-    "test_cases_count": 10,
-    "scores_baseline": [
-      -4.244843103350162,
-      -2.8540736410560337,
-      -3.5786205333621948,
-      -0.44690736170508316,
-      -3.5313951630134426,
-      -1.8214934191497276,
-      -3.2471992714871094,
-      -2.5140838715793548,
-      -1.4377568457998617,
-      -1.1753156509099216
-    ],
-    "scores_novel": [
-      6.255883736848207,
-      2.363713815798873,
-      0.9862179635835404,
-      2.7295724293178467,
-      -0.05232695425603795,
-      0.572151622917181,
-      1.4142691098214935,
-      2.3723483410121875,
-      3.9896976484255022,
-      4.220161147944019
-    ]
-  }
-}
