+++ tests/assays/test_plugin/test_integration_bradleyterryevaluator.json
