+++ tests/assays/test_plugin/test_integration_bradleyterryevaluator.readout.json
@@ -1,0 +1,30 @@
+{
+  "passed": true,
+  "details": {
+    "test_cases_count": 10,
+    "scores_baseline": [
+      -1.4070741733438608,
+      0.9335212155473751,
+      -1.1361640563376683,
+      2.2343024619532965,
+      -1.0969146982732503,
+      1.9243691482078238,
+      -2.9568837005600166,
+      0.9395660650820212,
+      -2.402725773124154,
+      0.035237535574653256
+    ],
+    "scores_novel": [
+      -0.8655849785739451,
+      -2.464926901043731,
+      -2.077590619536874,
+      2.3090948664078192,
+      1.5623501391714452,
+      1.204159655664255,
+      1.489381151479081,
+      -0.5109084129037269,
+      -0.5911641367838494,
+      2.877955211393285
+    ]
+  }
+}
