+++ tests/assays/test_plugin/test_integration_pairwiseevaluator.readout.json
